Segunda Iteración (reverted from commit 8bd4eec8b1afd377c8b1e41be4e2d8e265f08ae0)

+++ src/main/java/ar/com/ddsutn/TP1Macowins/Camisa.java
@@ -3,7 +3,7 @@ package ar.com.ddsutn.TP1Macowins;
 public class Camisa extends Prenda{
 	
 	@Override
-	protected double precioBase(){
+	protected int precioBase(){
 		return 200;
 	}
